requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv
httpx>=0.27
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
tenacity>=8.0
pytest
pytest-asyncio

